requirements.txt
# Core Framework
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.5

# LangChain Core
langchain==1.1.0
langchain-core==1.1.0
langchain-community==0.4.1
langgraph==1.0.4

# LLM Providers
langchain-google-genai==3.2.0
langchain-groq==1.1.0

# Search Tools
langchain-tavily==0.2.13
duckduckgo-search==8.1.1
ddgs==9.9.2

# HTTP & Networking
httpx==0.28.1
aiohttp==3.13.2

# Utilities
tenacity==9.1.2
